UniversalPayload: Update BootloaderParameter fetch logic
DO NOT MERGE: THIS IS A WA

Current UniversalPayload is designed as EDK2 payload, initialized
environment would be set up by EDK2 and parameter which will be
passed to UniversalPayload is provided by EDK2.

However, if UniversalPayload is not loaded by EDK2, parameter
might not be provided. In this case, BootloaderParameter should
be fetched in other method. In this WA, BootloaderParameter will
be used as Hob list, and it will be fetched 1MB ahead of payload
FD base.

Signed-off-by: Ajan Zhong <ajan.zhong@newfw.com>